% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_at.R
\name{ggplot_at}
\alias{ggplot_at}
\title{The ggplot2 equivalent to the scoped tidyverse functions.}
\usage{
ggplot_at(data, xvars, yvars, limit = 100)
}
\arguments{
\item{data}{Default dataset to use for plot.}

\item{xvars}{Variables to include/exclude in the plot. You can use same specifications as in \code{\link[dplyr]{select}}.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
The ggplot2 equivalent to the scoped tidyverse functions.
}
\examples{
ggplot_at(mtcars, vars(mpg:disp), vars(hp:wt), limit = 9)

ggplot_at(mtcars, vars(mpg:disp), vars(hp:wt), limit = 9) \%>\%
 add_layer(geom_point())

ggplot_at(mtcars, vars(mpg:disp), vars(hp:wt), limit = 9) \%>\%
 add_layer(geom_point(), aes(color = qsec))
}
